pyproject.toml
[tool.poetry]
name = "turbulette"
version = "0.4.0"
description = "A batteries-included framework to build high performance, async GraphQL APIs"
license = "BSD-3-Clause"
authors = ["Matthieu MN <matthieu.macnab@pm.me>"]
maintainers = ["Matthieu MN <matthieu.macnab@pm.me>"]
homepage = "https://python-turbulette.github.io/turbulette/"
documentation = "https://python-turbulette.github.io/turbulette/"
repository = "https://github.com/python-turbulette/turbulette/"
keywords = ["grapql", "api", "asyncio", "gino", "ariadne"]
readme = "README.md"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Natural Language :: English",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Typing :: Typed",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[tool.poetry.urls]
"Ariadne" = "https://ariadnegraphql.org/"
"GINO" = "https://python-gino.org/docs/en/master/index.html"
"Starlette" = "https://www.starlette.io/"
"Pydantic" = "https://pydantic-docs.helpmanual.io/"
"Alembic" = "https://alembic.sqlalchemy.org/en/latest/"

[tool.poetry.dependencies]
python = "^3.6"
ariadne = ">=0.11,<0.13"
alembic = "^1.4.2"
pydantic = { extras = ["email"], version = "^1.6.1" }
passlib = { extras = ["bcrypt"], version = "^1.7.2" }
simple-settings = ">=0.19.1,<1.1.0"
python-jwt = "^3.2.6"
gino = { extras = ["starlette"], version = "^1.0.1" }
click = "^7.1.2"
psycopg2 = "^2.8.5"
async-caches = "^0.3.0"
ciso8601 = "^2.1.3"
argon2-cffi = { version = "^20.1.0", optional = true }

# dev extras
mkdocs-material = { version = "^6.0.1", optional = true }
mkdocs-git-revision-date-plugin = { version = "^0.3", optional = true }
mkdocstrings = {version = "^0.14.0", optional = true}
tuna = { version = "^0.5.0", optional = true }
memory_profiler = { version = "^0.58.0", optional = true }
pygments-graphql-lexer = { git = "https://github.com/gazorby/pygments-graphql-lexer.git", optional = true }
pymdown-extensions = {version = "^8.0.1", optional = true}


[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"
uvicorn = {extras = ["watchgodreload"], version = "^0.12.1"}
pytest = "^6.2.1"
pytest-cov = "^2.11.1"
pytest-asyncio = "^0.14.0"
coverage = {extras = ["toml"], version = "^5.3"}
async-asgi-testclient = "^1.4.5"
python-multipart = "^0.0.5"
mypy = "^0.790"
debugpy = "^1.2.0"

[tool.poetry.extras]
dev_doc = [
  "mkdocs-material",
  "mkdocs-git-revision-date-plugin",
  "mkdocstrings",
  "pygments-graphql-lexer",
  "pymdown-extensions",
]
dev_profiling = ["tuna", "memory_profiler"]
argon2 = ["argon2-cffi"]

[tool.poetry.plugins."pytest11"]
"turbulette" = "turbulette.test.pytest_plugin"

[tool.poetry.scripts]
"turb" = "turbulette.management.cli:cli"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov --cov-report html --cov-report xml"

[tool.coverage.run]
omit = ["turbulette/test/pytest_plugin.py"]
source = ["turbulette/"]

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '(?<!get-poetry)\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | snapshots
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
